Move Key out of GenericMap.h.
