+++ .gitignore
@@ -142,7 +142,7 @@ tests/test_cameraSys
 tests/test_coord
 tests/test_endpoint
 tests/test_functorKeys
-tests/test_genericmap
+tests/test_genericmapkey
 tests/test_imageHash
 tests/test_key
 tests/test_polygon
